Create UI for adjusting dashboard column width
Users with accessibility concerns can either tab onto the slider and
move it with the arrow keys, or tab into one of the width text boxes and
enter the number of columns manually.

This only appears on the dashboard configuration page, and assumes that
dashboards will be two columns wide.