Explicitly save submitted time args
For cases where pages are reloaded for a user, like adding
attachments on create, we want to restore all settings,
including the time units. Previously this worked because the
EditTimeValue looked directly at the original request_args.

With the new redirect, the originally submitted args aren't there,
so stash them explicitly in a new arg.